Moved the background image helper function to utils
